+++ src/main/java/pl/krepec/stockExchange/service/UserService.java
@@ -52,7 +52,7 @@ public class UserService {
         return null;
     }
 
-    public String updateUserCash(Integer id, Double quantity, Operation operation, String stockSymbol) {
+    public String updateUserCash(Integer id, Integer quantity, Operation operation, String stockSymbol) {
         UserDAO userDAO = userRepository.findOne(id);
         Double cashAfterShopping = market.shopping(id, quantity, operation, stockSymbol);
         userDAO.setCash(cashAfterShopping);
